requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.3
python-multipart==0.0.20
openai==1.59.3
anthropic==0.42.0
google-generativeai==0.8.3
cohere==5.13.4
requests==2.32.3
twilio==9.4.0
PyJWT==2.10.1
PyPDF2==3.0.1
pandas==2.2.3
python-docx==1.1.2
